NF: allow for sign flip in image comparison
The statistic images include tests on the second of two basis functions
from the spectral set.  This has a more or less random sign. About a
year ago I (MB) put a constraint on the sign (after discussing with
Jonathan) so the sign need not be the same as the previous run of this
analysis.